Made LevelManager responsible for pausing and unpausing
