+++ Assets/Scripts/Game Managers/LevelManager.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using UnityEngine.UI;
 
 public class LevelManager : MonoBehaviour
 {
@@ -9,20 +10,33 @@ public class LevelManager : MonoBehaviour
 
 	public Scene currentScene;
 
+    public Canvas PauseScreen;
+
     private PlayerController player;
 
+    private bool isPaused = false;
 
-
+    public float TimeScale = 1;
 
 	// Use this for initialization
 	void Start ()
     {
-        player = FindObjectOfType<PlayerController>();		
+        player = FindObjectOfType<PlayerController>();
+        PauseScreen.gameObject.SetActive(false);//Hide pause screen		
 	}
 	
 	// Update is called once per frame
 	void Update () {
-		
+		if(Input.GetButtonDown("Pause"))
+        {
+            if (isPaused)
+            {
+                Unpause();
+            } else
+            {
+                Pause();
+            }
+        }
 	}
 
     public void RespawnPlayer ()
@@ -38,5 +52,23 @@ public class LevelManager : MonoBehaviour
             Cursor.lockState = CursorLockMode.None;
             SceneManager.LoadScene ("MainMenu");
 		}
+    }
+
+    public void Pause()
+    {
+        isPaused = true;
+        Time.timeScale = 0;
+        Cursor.lockState = CursorLockMode.None;
+        PauseScreen.gameObject.SetActive(true);
+        Debug.Log("pause");
+    }
+
+    public void Unpause()
+    {
+        isPaused = false;
+        Time.timeScale = TimeScale;
+        Cursor.lockState = CursorLockMode.Locked;
+        PauseScreen.gameObject.SetActive(false);
+        Debug.Log("unpause");
     }
 }
