--- Assets/Scripts/Game Managers/PauseScreen.cs
@@ -1,94 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class PauseScreen : MonoBehaviour
-{
-    //TODO move this all to LevelManager
-    private GameObject[] pauseObjects;
-    //TODO instead hold a canvas with all the pause stuff on it, and activate/deactivate that
-
-    // Use this for initialization
-    void Start()
-    {
-        Time.timeScale = 1;
-        pauseObjects = GameObject.FindGameObjectsWithTag("ShowOnPause");
-        hidePaused();
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        //TODO change to use Pause Button axis
-        //TODO variable to remember if we're paused, don't just rely on timescale
-        //uses the p button to pause and unpause the game
-        if (Input.GetKeyDown(KeyCode.P))
-        {
-            if (Time.timeScale == 1)
-            {
-                Time.timeScale = 0;
-                Cursor.lockState = CursorLockMode.None;
-                showPaused();
-            }
-            else if (Time.timeScale == 0)
-            {
-                Debug.Log("high");
-                Time.timeScale = 1;
-                Cursor.lockState = CursorLockMode.Locked;
-                hidePaused();
-            }
-        }
-    }
-
-
-    //Reloads the Level
-    public void Reload()
-    {
-        Application.LoadLevel(Application.loadedLevel);
-    }
-
-    public void ReturnToMenu()
-    {
-        SceneManager.LoadScene("MainMenu");
-    }
-
-    //controls the pausing of the scene
-    public void pauseControl()
-    {
-        if (Time.timeScale == 1)
-        {
-            Time.timeScale = 0;
-            showPaused();
-        }
-        else if (Time.timeScale == 0)
-        {
-            Time.timeScale = 1;
-            hidePaused();
-        }
-    }
-
-    //shows objects with ShowOnPause tag
-    public void showPaused()
-    {
-        foreach (GameObject g in pauseObjects)
-        {
-            g.SetActive(true);
-        }
-    }
-
-    //hides objects with ShowOnPause tag
-    public void hidePaused()
-    {
-        foreach (GameObject g in pauseObjects)
-        {
-            g.SetActive(false);
-        }
-    }
-
-    //loads inputted level
-    public void LoadLevel(string level)
-    {
-        Application.LoadLevel(level);
-    }
-}
